# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "DotMac Platform Services API",
    "version": "1.0.0",
    "description": "Canonical OpenAPI contract used by integration tests to validate core authentication and platform service endpoints."
  },
  "servers": [
    {
      "url": "https://api.dotmac.local",
      "description": "Local testing server"
    }
  ],
  "paths": {
...
    "/api/v1/deployment/stats": {
      "get": {
        "summary": "Get deployment statistics summary",
        "tags": [
          "Deployment"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "tenant_id",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "description": "Filter statistics for a specific tenant"
          }
        ],
        "responses": {
          "200": {
            "description": "Aggregated deployment statistics",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeploymentStatsResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/deployment/stats/resources": {
      "get": {
        "summary": "Get resource allocation totals",
        "tags": [
          "Deployment"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "tenant_id",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "description": "Filter allocation totals for a specific tenant"
          }
        ],
        "responses": {
          "200": {
            "description": "Aggregated resource allocation totals",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeploymentResourceAllocationResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
...
    "schemas": {
...
      "DeploymentStatsResponse": {
        "type": "object",
        "required": [
          "total_instances",
          "states",
          "health"
        ],
        "properties": {
          "total_instances": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "states": {
            "type": "object",
            "required": [
              "active",
              "provisioning",
              "failed",
              "suspended"
            ],
            "properties": {
              "active": {
                "type": "integer",
                "format": "int64",
                "minimum": 0
              },
              "provisioning": {
                "type": "integer",
                "format": "int64",
                "minimum": 0
              },
              "failed": {
                "type": "integer",
                "format": "int64",
                "minimum": 0
              },
              "suspended": {
                "type": "integer",
                "format": "int64",
                "minimum": 0
              }
            }
          },
          "health": {
            "type": "object",
            "required": [
              "healthy",
              "degraded",
              "unhealthy"
            ],
            "properties": {
              "healthy": {
                "type": "integer",
                "format": "int64",
                "minimum": 0
              },
              "degraded": {
                "type": "integer",
                "format": "int64",
                "minimum": 0
              },
              "unhealthy": {
                "type": "integer",
                "format": "int64",
                "minimum": 0
              }
            }
          }
        }
      },
...
      "DeploymentResourceAllocationResponse": {
        "type": "object",
        "required": [
          "total_cpu",
          "total_memory",
          "total_storage"
        ],
        "properties": {
          "total_cpu": {
            "type": "integer",
            "format": "int64",
            "minimum": 0,
            "description": "Total allocated CPU cores"
          },
          "total_memory": {
            "type": "integer",
            "format": "int64",
            "minimum": 0,
            "description": "Total allocated memory in GiB"
          },
          "total_storage": {
            "type": "integer",
            "format": "int64",
            "minimum": 0,
            "description": "Total allocated storage in GiB"
          }
        }
      }
    }
  }
}